refactor(#3479): extract task-notification lifecycle from turn_bridge/mod.rs
Verbatim move of merge_task_notification_kind, release_task_notification_kind,
close_next_tracked_background_child, close_all_tracked_background_children,
task_notification_closes_background_child + their lifecycle test module into
new sibling turn_bridge/task_notification_lifecycle.rs (105 prod / 117 test LoC),
re-exported pub(super). turn_bridge/mod.rs prod LoC 6535 -> 6448; giant ratchet
baseline lowered to 6448.
